Split out normalizePostForEditing
